+++ packages/compass-ui/src/components/ModalFooter/index.ts
@@ -1,0 +1,2 @@
+export { default as ModalFooter } from './ModalFooter';
+export type { ModalFooterProps, ModalFooterType } from './ModalFooter';